+++ docs/release-notes/change-log.md
@@ -17,6 +17,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Fix handling of 'special case' substreams module with only "params" as its input: should not skip this execution (used in graph-node for head tracking) 
   -> empty files in module cache with hash `d3b1920483180cbcd2fd10abcabbee431146f4c8` should be deleted for consistency
 
+### CLI
+
+* Add `substreams tools default-endpoint {network-name}` to help with auto-configuration tools
+
 ## v1.10.2
 
 * `substreams gui`: fix panic in some conditions when streaming from block 0